Added volumes: parameter to task
when implementing the pipelinerun through the UI only one pvc can be
associated to the resource. If we were applying the pipelinerun through
a trigger or manually (oc create -f) this would be a moot point.
However, since we are using the interface for the demo it makes sense to
add the volumes: section to the task file which allows for separate
storage devices for the pipelines and containers.