Switch to using global.localClusterDomain for StackRox clusterName
Instead of adding a new localClusterName variable which is quite a bit
of work due to a number of reasons (it needs to be set via an ACM policy
too and there is no split support with strings until ACM 2.6, so it
becomes hard to extract it). Let's just use localClusterDomain and take
out the cluster domain from there.

Tested this and went over the result with Jonny. ACS shows the hub
cluster with its DNS name, and importing a remote cluster via ACM
correctly gets us the cluster in ACS as well.