Common (#100)
* Add some comments to make if/else and loops clearer

Let's improve readability by adding some comments to point out which
flow constructs are being ended.

* Add some more comments in applications.yaml

* Add a default for options applicationRetryLimit

* Split out values files to a helper for the acm chart

Just like we did for the clustergroup chart, let's split the values
file list into a dedicated helper. This time since there are no global
variables we include it with the current context and not with the '$'
context.

Tested with MCG: hub and spoke. Correctly observed all the applications
running on the spoke.

* Fix up tests

They changed because we made the list indentation more correct (two
extra spaces to the left)

* Fix sa/namespace mixup in vault_spokes_init

* Update local patch

Also set seccompProfile to null to make things work on OCP 4.10

* Update ESO to 0.8.5

* Tweak ESO UBI images

Tested the ESO upgrade on MCG on both 4.10 and 4.13

* Removed previous version of common to convert to subtree from https://github.com/hybrid-cloud-patterns/common.git main

* make test

---------

Co-authored-by: Michele Baldessari <michele@acksyn.org>
Co-authored-by: Tom Stockwell <2060486+stocky37@users.noreply.github.com>